requirements.txt
langchain
langchain-community
langchain-openai
anthropic
chromadb
pypdf
unstructured
python-dotenv
tiktoken
rich
rouge-score
sacrebleu
bert-score
sentence-transformers
# Retrieval
rank-bm25
# Phase 2: Accessibility - Speech-to-Text
openai-whisper
sounddevice
numpy
scipy
# Phase 2: Accessibility - Text-to-Speech
edge-tts
# pyttsx3  # Replaced with edge-tts for better quality
rank_bm25
